requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.2
oauthlib==3.1.1
psycopg2-binary==2.9.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
